+++ changelog.md
@@ -12,6 +12,8 @@ This project attempts to adheres to [Semantic Versioning](http://semver.org/).
 - Iterables.filter returning a new iterable containing only those elements for which the predicate returns true
 - Iterables.join transforms a nested iterable of iterables into a single iterable
 - Iterables.addAll adds all of the input collection into the passed iterable
+- Iterables.cycle returns an infinite iterable that cycles through the input elements (does not support removing elements)
+- Iterables.makeString pretty prints the contents of an iterable
 - Iterables.concat creates a single iterable containing all of the input iterables values
 - Options.nullSafe transforms a null producing function into one returning an option
 - Options.toOption returns a function that builds an option